Adicionado Crud e relatório

+++ database/migrations/2023_12_20_175733_create_livros_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('editora', 40);
             $table->integer('edicao');
             $table->string('ano_publicacao', 4);
-            $table->float('preco');
+            $table->decimal('preco', 20);
             $table->timestamps();
         });
     }
